[executor] remove execution_service, execution_proto and execution_client
