--- execution/execution_client/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "execution_client"
-version = "0.1.0"
-authors = ["Libra Association <opensource@libra.org>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-grpcio = { version = "0.4.4", default-features = false, features = ["protobuf-codec"] }
-
-execution_proto = { path = "../execution_proto" }
-failure = { path = "../../common/failure_ext", package = "failure_ext" }
-crypto = { path = "../../crypto/crypto" }
-types = { path = "../../types" }
-proto_conv = { path = "../../common/proto_conv" }
-
-[dev-dependencies]
-types = { path = "../../types", features = ["testing"] }
