--- execution/execution_proto/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "execution_proto"
-version = "0.1.0"
-authors = ["Libra Association <opensource@libra.org>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-futures = "0.1.28"
-grpcio = { version = "0.4.4", default-features = false, features = ["protobuf-codec"] }
-proptest = { version = "0.9.4", default-features = false }
-proptest-derive = { version = "0.1.2", default-features = false }
-protobuf = "~2.7"
-
-crypto = { path = "../../crypto/crypto" }
-failure = { path = "../../common/failure_ext", package = "failure_ext" }
-proptest_helpers = { path = "../../common/proptest_helpers" }
-proto_conv = { path = "../../common/proto_conv", features = ["derive"] }
-types = { path = "../../types" }
-
-[build-dependencies]
-build_helpers = { path = "../../common/build_helpers" }
-
-[dev-dependencies]
-types = { path = "../../types", features = ["testing"] }
