--- execution/execution_proto/src/proto/mod.rs
@@ -1,9 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-#![allow(bare_trait_objects)]
-
-use types::proto::*;
-
-pub mod execution;
-pub mod execution_grpc;
