--- execution/execution_service/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "execution_service"
-version = "0.1.0"
-authors = ["Libra Association <opensource@libra.org>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-futures01 = { version = "0.1.28", default-features = false }
-futures03 = { package = "futures-preview", version = "=0.3.0-alpha.17", features = ["compat"] }
-grpcio = { version = "0.4.4", default-features = false }
-
-config = { path = "../../config" }
-execution_proto = { path = "../execution_proto" }
-executor = { path = "../executor" }
-failure = { path = "../../common/failure_ext", package = "failure_ext" }
-grpc_helpers = { path = "../../common/grpc_helpers" }
-proto_conv = { path = "../../common/proto_conv" }
-storage_client = { path = "../../storage/storage_client" }
-vm_runtime = { path = "../../language/vm/vm_runtime" }
-
-[dev-dependencies]
-types = { path = "../../types", features = ["testing"] }
