--- execution/execution_tests/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "execution_tests"
-version = "0.1.0"
-authors = ["Libra Association <opensource@libra.org>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-futures01 = { version = "0.1.28", default-features = false }
-grpcio = { version = "0.4.4", default-features = false }
-
-config = { path = "../../config" }
-config_builder = { path = "../../config/config_builder" }
-crypto = { path = "../../crypto/crypto" }
-execution_client = { path = "../execution_client" }
-execution_proto = { path = "../execution_proto" }
-execution_service = { path = "../execution_service" }
-failure = { path = "../../common/failure_ext", package = "failure_ext" }
-grpc_helpers = { path = "../../common/grpc_helpers" }
-proto_conv = { path = "../../common/proto_conv" }
-storage_client = { path = "../../storage/storage_client" }
-storage_service = { path = "../../storage/storage_service" }
-types = { path = "../../types" }
-vm_genesis = { path = "../../language/vm/vm_genesis" }
-
-[dev-dependencies]
-rand = "0.6.5"
-crypto = { path = "../../crypto/crypto", features = ["testing"] }
-types = { path = "../../types", features = ["testing"]}
