--- execution/execution_tests/src/tests/mod.rs
@@ -1,7 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-// Add new test modules here.
-
-mod execution_service_test;
-mod storage_integration_test;
