+++ execution/executor/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 backoff = { version = "0.1.5", default-features = false }
+config_builder = { path = "../../config/config_builder" }
 futures = { version = "=0.3.0-alpha.17", package = "futures-preview" }
 itertools = { version = "0.8.0", default-features = false }
 lazy_static = { version = "1.3.0", default-features = false }
@@ -17,7 +18,6 @@ serde = { version = "1.0.99", default-features = false }
 canonical_serialization = { path = "../../common/canonical_serialization" }
 config = { path = "../../config" }
 crypto = { path = "../../crypto/crypto" }
-execution_proto = { path = "../execution_proto" }
 failure = { path = "../../common/failure_ext", package = "failure_ext" }
 logger = { path = "../../common/logger" }
 metrics = { path = "../../common/metrics" }
@@ -32,9 +32,11 @@ vm_genesis = { path = "../../language/vm/vm_genesis" }
 [dev-dependencies]
 grpcio = { version = "0.4.4", default-features = false, features = ["protobuf-codec"] }
 proptest = "0.9.2"
+rand = "0.6.5"
 rusty-fork = "0.2.1"
 
 config = { path = "../../config", features = ["testing"]}
+grpc_helpers = { path = "../../common/grpc_helpers" }
 types = { path = "../../types", features = ["testing"]}
 
 storage_proto = { path = "../../storage/storage_proto" }
